No more memory leak: list cleared

+++ Sources/Factory.cpp
@@ -47,7 +47,7 @@ namespace Abstract
     {
         if (std::stof(value) > std::numeric_limits<float>::max() || std::stof(value) < std::numeric_limits<float>::min())
             throw Exception("Type value overflow");
-        return new Operand<float>(std::stof(value), Float);
+        return new Operand<float>(std::stoi(value), Float);
     }
 
     IOperand* Factory::createDouble(const std::string &value)
